More Path changes, setting up for detection/clustering threshold dissociation
- Do_clustering:
A. plots now have names, and remain visible in the batch script.
B. the threshold variable "thr" is now stored with the parameters.
C. If a figure is empty, it is closed at the end.
D. Do_clustering also now accepts function handles as inputs.

-Plot_continuous_data
A. now passes out "handles" which it gets, which have the field "thr" added to "par" - this contains the threshold used for spike detection.

-parse_data_NSX
A. Creates a folder of the same name as the nsX file, in the same directory, with "_parsed" added, and puts the output of the processing there.
B. Restructured to return the filepath of the files parsed.
C. Restructured to first check if the directory it would make is already present, and if so, to return the filepaths of the files within.

Added "bulk_parse_data_NSX"
- takes in a few variables noted at the top, preprocesses the data found by the "buildRunList" function, which works with a path given the path at the top. Currently only really works for NSX.

Other smaller changes surrounding path management.